requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
adjusttext=0.7.3.1=py_1
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38_0
cffi=1.14.0=py38he30daa8_1
chardet=3.0.4=py38_1003
cpuonly=1.0=0
cryptography=2.9.2=py38h1ba5d50_0
cycler=0.10.0=py_2
freetype=2.9.1=h8a8886c_1
icu=58.2=he6710b0_3
idna=2.9=py_1
intel-openmp=2020.1=217
joblib=0.15.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py38hbf85e49_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libxml2=2.9.9=hea5a465_1
libxslt=1.1.33=h7d1a2b0_0
lxml=4.5.0=pypi_0
matplotlib=3.2.1=pypi_0
matplotlib-base=3.1.3=py38hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
multitasking=0.0.9=py_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py38hfd86e86_0
numpy=1.18.4=pypi_0
numpy-base=1.18.1=py38hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1g=h516909a_0
pandas=1.0.3=py38h0573a6f_0
pandas-datareader=0.8.1=py_0
pillow=7.1.2=py38hb39fc2d_0
pip=20.0.2=py38_3
pycparser=2.20=py_0
pyopenssl=19.1.0=py38_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38_0
python=3.8.2=hcff3b4d_14
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.0=py3.8_cpu_0
pytz=2020.1=py_0
readline=8.0=h7b6447c_0
requests=2.23.0=py38_0
requests-futures=1.0.0=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.5.0=pypi_0
selenium=3.141.0=pypi_0
setuptools=46.1.3=py38_0
six=1.14.0=py38_0
sqlite=3.31.1=h62c20be_1
threadpoolctl=2.1.0=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.6.0=py38_cpu
tornado=6.0.4=py38h1e0a361_1
urllib3=1.25.9=pypi_0
wheel=0.34.2=py38_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7b6447c_0
yahooquery=2.2.4=pypi_0
yfinance=0.1.54=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
